fixed more compiler warnings about: comparison between signed and unsigned integer expressions

+++ remeshing.cxx
@@ -146,7 +146,7 @@ void new_bottom(const uint_vec &old_bcflag, double *qcoord,
 
     // mark all bottom facets nodes as deleted
     for (int i=0; i<nseg; ++i) {
-        if (segflag[i] == BOUNDZ0) {
+        if (static_cast<uint>(segflag[i]) == BOUNDZ0) {
             for (int j=0; j<NODES_PER_FACET; j++)
                 segment[i*NODES_PER_FACET + j] = DELETED_FACET;
         }
@@ -159,7 +159,7 @@ void new_bottom(const uint_vec &old_bcflag, double *qcoord,
     //      |    |
     //      2 -- 3
     for (int i=0, nfacets=0, offset=0; i<nseg, nfacets<(NDIMS-1); ++i) {
-        if (segflag[i] == BOUNDZ0) {
+        if (static_cast<uint>(segflag[i]) == BOUNDZ0) {
             for (int j=0; j<NODES_PER_FACET; j++)
                 segment[i*NODES_PER_FACET + j] = bottom_corners[offset + j];
